+++ NEWS.md
@@ -4,6 +4,7 @@
 * Rename `nutrients_output` to `nutrients_loss`
 * Rename `detritus_output` to `detritus_loss`
 * Run all seagrass/water column processes on same temporal scale
+* Better global variables handling
 
 # arrR 1.3.1
 * Changes to `get_stable_values`